Add: IdentityServer, Authentification

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -53,14 +54,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
-            //TO-DO: Add current user as Seller
-
-            auction.Seller = "Test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -79,6 +79,8 @@ namespace AuctionService.Controllers
             );
         }
 
+
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -88,7 +90,7 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound("Auction not found");
 
-            // TO-DO Check seller username
+            if (auction.Seller != User.Identity.Name) return Forbid();            
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -106,6 +108,7 @@ namespace AuctionService.Controllers
 
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -113,7 +116,7 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound("Auction not found");
 
-            // TO-DO Check seller username
+            if (auction.Seller != User.Identity.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
